+++ src/calling/variants/preprocessing/mod.rs
@@ -75,7 +75,7 @@ pub(crate) struct ObservationProcessor<R: realignment::Realigner + Clone + 'stat
     raw_observation_output: Option<PathBuf>,
     report_fragment_ids: bool,
     adjust_prob_mapping: bool,
-    readtype: Readtype,
+    atomic_candidate_variants: bool,
 }
 
 impl<R: realignment::Realigner + Clone + std::marker::Send + std::marker::Sync>
@@ -505,6 +505,7 @@ impl<R: realignment::Realigner + Clone + std::marker::Send + std::marker::Sync>
                 ref_base()?,
                 alt,
                 self.realigner.clone(),
+                !self.atomic_candidate_variants,
             ))
         };
 
@@ -528,6 +529,7 @@ impl<R: realignment::Realigner + Clone + std::marker::Send + std::marker::Sync>
                     .to_owned(),
                 alt.to_owned(),
                 self.realigner.clone(),
+                !self.atomic_candidate_variants,
             ))
         };