+++ src/variants/evidence/realignment/pairhmm.rs
@@ -77,6 +77,8 @@ pub(crate) trait RefBaseEmission {
 
 pub(crate) trait VariantEmission {
     fn is_homopolymer_indel(&self) -> bool;
+
+    fn alt_vs_ref_len_diff(&self) -> isize;
 }
 
 #[macro_export]
@@ -367,6 +369,10 @@ impl<'a> VariantEmission for ReadVsAlleleEmission<'a> {
     fn is_homopolymer_indel(&self) -> bool {
         self.allele_emission.is_homopolymer_indel()
     }
+
+    fn alt_vs_ref_len_diff(&self) -> isize {
+        self.allele_emission.alt_vs_ref_len_diff()
+    }
 }
 
 impl<'a> pairhmm::EmissionParameters for ReadVsAlleleEmission<'a> {
@@ -519,6 +525,10 @@ impl RefBaseEmission for ReferenceEmissionParams {
 impl VariantEmission for ReferenceEmissionParams {
     fn is_homopolymer_indel(&self) -> bool {
         false
+    }
+
+    fn alt_vs_ref_len_diff(&self) -> isize {
+        0
     }
 }